refactor(actividades): separar queries de controlador para simplificar estructura
