+++ models/otros/anuncio.model.js
@@ -52,19 +52,15 @@ async function getAnnouncements(){
 }
 
 async function patchAnnouncement(IDAnuncio, titulo, contenido, imagen){
-    try {
-        const announcement = await Anuncio.findById(IDAnuncio);
-        if (announcement) {
-            announcement.titulo = titulo;
-            announcement.contenido = contenido;
-            announcement.imagen = imagen;
-            await announcement.save();
-            return announcement;
-        } else {
-            throw new Error('Anuncio no encontrado');
-        }
-    } catch (error) {
-        throw error;
+    const announcement = await Anuncio.findById(IDAnuncio);
+    if (announcement) {
+        announcement.titulo = titulo;
+        announcement.contenido = contenido;
+        announcement.imagen = imagen;
+        await announcement.save();
+        return announcement;
+    } else {
+        throw new Error('Anuncio no encontrado');
     }
 }
