+++ package.json
@@ -34,6 +34,7 @@
     "moment-timezone": "^0.5.46",
     "mongoose": "^8.6.0",
     "multer": "^1.4.5-lts.1",
+    "node-cron": "^3.0.3",
     "nodemon": "^3.1.4"
   },
   "devDependencies": {
